+++ app/src/main/java/com/example/finances/domain/models/ShopItemPrice.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public class ShopItemPrice extends Price {
     private ShopItemPriceType additionalType;
 
-    public ShopItemPrice(int id, float price, LocalDate created, LocalDate modified, PriceType type, ShopItemPriceType additionalType) {
+    public ShopItemPrice(long id, float price, LocalDate created, LocalDate modified, PriceType type, ShopItemPriceType additionalType) {
         super(id, price, created, modified, type);
 
         this.additionalType = additionalType;
